refactor: Updated webframework-shim to reduce the cognitive complexity
