+++ lib/shim/webframework-shim.js
@@ -5,8 +5,6 @@
 
 'use strict'
 
-/* eslint sonarjs/cognitive-complexity: ["error", 69] -- TODO: https://issues.newrelic.com/browse/NEWRELIC-5252 */
-
 const genericRecorder = require('../metrics/recorders/generic')
 const logger = require('../logger.js').child({ component: 'WebFrameworkShim' })
 const metrics = require('../metrics/names')
@@ -14,8 +12,15 @@ const TransactionShim = require('./transaction-shim')
 const Shim = require('./shim')
 const specs = require('./specs')
 const util = require('util')
-const symbols = require('../symbols')
-const { assignCLMSymbol } = require('../util/code-level-metrics')
+const {
+  assignError,
+  copyExpectedSpecParameters,
+  getTransactionInfo,
+  isError,
+  MIDDLEWARE_TYPE_NAMES
+} = require('./webframework-shim/common')
+const wrapMiddlewareMounter = require('./webframework-shim/middleware-mounter')
+const { _recordMiddleware } = require('./webframework-shim/middleware')
 
 /**
  * An enumeration of well-known web frameworks so that new instrumentations can
@@ -40,24 +45,6 @@ const FRAMEWORK_NAMES = {
   RESTIFY: 'Restify'
 }
 
-const MIDDLEWARE_TYPE_DETAILS = {
-  APPLICATION: { name: 'Mounted App: ', path: true, record: false },
-  ERRORWARE: { name: '', path: false, record: true },
-  MIDDLEWARE: { name: '', path: false, record: true },
-  PARAMWARE: { name: '', path: false, record: true },
-  ROUTE: { name: 'Route Path: ', path: true, record: false },
-  ROUTER: { name: 'Router: ', path: true, record: false }
-}
-
-const MIDDLEWARE_TYPE_NAMES = {
-  APPLICATION: 'APPLICATION',
-  ERRORWARE: 'ERRORWARE',
-  MIDDLEWARE: 'MIDDLEWARE',
-  PARAMWARE: 'PARAMWARE',
-  ROUTE: 'ROUTE',
-  ROUTER: 'ROUTER'
-}
-
 /**
  * Constructs a shim associated with the given agent instance, specialized for
  * instrumenting web frameworks.
@@ -68,7 +55,6 @@ const MIDDLEWARE_TYPE_NAMES = {
  *  A helper class for wrapping web framework modules.
  * @param {Agent} agent
  *  The agent this shim will use.
- * @param shimName
  * @param {string} moduleName
  *  The name of the module being instrumented.
  * @param {string} resolvedName
@@ -153,7 +139,6 @@ WebFrameworkShim.prototype.savePossibleTransactionName = savePossibleTransaction
  * @param {Function}            fn      - The middleware function.
  * @param {string}              fnName  - The name of the middleware function.
  * @param {Array}               args    - The arguments to the middleware function.
- * @param {NextWrapperFunction} wrap    - A function to wrap an individual next function.
  * @returns {object} The request object.
  */
 
@@ -203,7 +188,7 @@ WebFrameworkShim.prototype.savePossibleTransactionName = savePossibleTransaction
  * @property {RouteParameterFunction} [params]
  *  A function to extract the route parameters from the middleware arguments.
  *  Defaults to using `req.params`.
- * @property {string} [type='MIDDLEWARE']
+ * @property {string} [type='MIDDLEWARE'] specifies the type
  * @property {string | Function} [route=null]
  *  Route/path used for naming segments and transaction name candidates. If a function,
  *  will be invoked just before segment creation with middleware invocation.
@@ -247,10 +232,12 @@ WebFrameworkShim.prototype.savePossibleTransactionName = savePossibleTransaction
  *
  * @memberof WebFrameworkShim.prototype
  * @param {RouteParserFunction} parser - The parser function to use.
+ * @returns {undefined}
  */
 function setRouteParser(parser) {
   if (!this.isFunction(parser)) {
-    return this.logger.debug('Given route parser is not a function.')
+    this.logger.debug('Given route parser is not a function.')
+    return
   }
   this._routeParser = parser
 }
@@ -346,111 +333,6 @@ function recordRender(nodule, properties, spec) {
 }
 
 /**
- * Wraps a method that is used to add middleware to a server. The middleware
- * can then be recorded as metrics.
- *
- * - `wrapMiddlewareMounter(nodule, properties [, spec])`
- * - `wrapMiddlewareMounter(func [, spec])`
- *
- * @memberof WebFrameworkShim.prototype
- * @param {object | Function} nodule
- *  The source for the properties to wrap, or a single function to wrap.
- * @param {string|Array.<string>} [properties]
- *  One or more properties to wrap. If omitted, the `nodule` parameter is
- *  assumed to be the function to wrap.
- * @param {MiddlewareMounterSpec} [spec]
- *  Spec describing the parameters for this middleware mount point.
- * @returns {object | Function} The first parameter to this function, after
- *  wrapping it or its properties.
- * @see WebFrameworkShim#recordMiddleware
- */
-function wrapMiddlewareMounter(nodule, properties, spec) {
-  if (properties && !this.isString(properties) && !this.isArray(properties)) {
-    // wrapMiddlewareMounter(func, spec)
-    spec = properties
-    properties = null
-  }
-  if (this.isFunction(spec)) {
-    // wrapMiddlewareMounter(nodule [, properties], wrapper)
-    spec = { wrapper: spec }
-  }
-
-  spec = this.setDefaults(spec, {
-    route: null,
-    endpoint: null
-  })
-
-  const wrapSpec = {
-    wrapper: function wrapMounter(shim, fn, fnName) {
-      if (!shim.isFunction(fn)) {
-        return fn
-      }
-
-      return function wrappedMounter() {
-        const args = shim.argsToArray.apply(shim, arguments)
-
-        // Normalize the route index and pull out the route argument if provided.
-        let routeIdx = null
-        let route = null
-        if (shim.isNumber(spec.route)) {
-          routeIdx = shim.normalizeIndex(args.length, spec.route)
-          route = routeIdx === null ? null : args[routeIdx]
-          const isArrayOfFunctions = shim.isArray(route) && shim.isFunction(route[0])
-          if (shim.isFunction(route) || isArrayOfFunctions) {
-            routeIdx = null
-            route = null
-          } else if (shim.isArray(route)) {
-            route = route.map((routeArg) => {
-              return shim._routeParser.call(this, shim, fn, fnName, routeArg)
-            })
-          } else {
-            route = shim._routeParser.call(this, shim, fn, fnName, route)
-          }
-        } else if (spec.route !== null) {
-          route = shim._routeParser.call(this, shim, fn, fnName, spec.route)
-        }
-
-        _wrapMiddlewares.call(this, routeIdx, args)
-        /**
-         * @param _routeIdx
-         * @param middlewares
-         */
-        function _wrapMiddlewares(_routeIdx, middlewares) {
-          for (let i = 0; i < middlewares.length; ++i) {
-            // If this argument is the route argument skip it.
-            if (i === _routeIdx) {
-              continue
-            }
-
-            // Some platforms accept an arbitrarily nested array of middlewares,
-            // so if this argument is an array we must recurse into it.
-            const middleware = middlewares[i]
-            if (middleware instanceof Array) {
-              _wrapMiddlewares(null, middleware)
-              continue
-            }
-
-            middlewares[i] = spec.wrapper.call(
-              this,
-              shim,
-              middleware,
-              shim.getName(middleware),
-              route
-            )
-          }
-        }
-
-        return fn.apply(this, args)
-      }
-    }
-  }
-
-  _copyExpectedSpecParameters(wrapSpec, spec)
-
-  return this.wrap(nodule, properties, wrapSpec)
-}
-
-/**
  * Records the provided function as a middleware.
  *
  * - `recordMiddleware(nodule, properties [, spec])`
@@ -481,7 +363,7 @@ function recordMiddleware(nodule, properties, spec) {
     return _recordMiddleware(shim, middleware, mwSpec)
   })
 
-  _copyExpectedSpecParameters(wrapSpec, spec)
+  copyExpectedSpecParameters(wrapSpec, spec)
 
   return this.wrap(nodule, properties, wrapSpec)
 }
@@ -531,7 +413,7 @@ function recordParamware(nodule, properties, spec) {
     return _recordMiddleware(shim, middleware, mwSpec)
   })
 
-  _copyExpectedSpecParameters(wrapSpec, spec)
+  copyExpectedSpecParameters(wrapSpec, spec)
 
   return this.wrap(nodule, properties, wrapSpec)
 }
@@ -552,9 +434,9 @@ function recordParamware(nodule, properties, spec) {
  * @see WebFrameworkShim#setErrorPredicate
  */
 function noticeError(req, err) {
-  const txInfo = _getTransactionInfo(this, req)
-  if (txInfo && _isError(this, err)) {
-    _noticeError(this, txInfo, err)
+  const txInfo = getTransactionInfo(this, req)
+  if (txInfo && isError(this, err)) {
+    assignError(txInfo, err)
   }
 }
 
@@ -568,7 +450,7 @@ function noticeError(req, err) {
  * @see WebFrameworkShim#setErrorPredicate
  */
 function errorHandled(req, err) {
-  const txInfo = _getTransactionInfo(this, req)
+  const txInfo = getTransactionInfo(this, req)
   if (txInfo && txInfo.error === err) {
     txInfo.errorHandled = true
   }
@@ -579,7 +461,7 @@ function errorHandled(req, err) {
  * an error.
  *
  * @memberof WebFrameworkShim.prototype
- * @param {function(object): bool} pred
+ * @param {function(object): boolean} pred
  *  Function which should return true if the object passed to it is considered
  *  an error.
  * @see WebFrameworkShim#noticeError
@@ -596,7 +478,7 @@ function setErrorPredicate(pred) {
  * @param {Request} req - The request which caused the error.
  */
 function savePossibleTransactionName(req) {
-  const txInfo = _getTransactionInfo(this, req)
+  const txInfo = getTransactionInfo(this, req)
   if (txInfo && txInfo.transaction) {
     txInfo.transaction.nameState.markPath()
   }
@@ -607,7 +489,7 @@ function savePossibleTransactionName(req) {
  * responded.
  *
  * @memberof WebFrameworkShim.prototype
- * @param {function(args, object): bool} pred
+ * @param {function(args, object): boolean} pred
  *  Function which should return true if the object passed to it is considered
  *  a response.
  */
@@ -629,6 +511,7 @@ function setResponsePredicate(pred) {
  *  The name of the function to which this route was given.
  * @param {string|RegExp} route
  *  The route that was given to the function.
+ *  @returns {string} route name
  * @see RouteParserFunction
  */
 function _defaultRouteParser(shim, fn, fnName, route) {
@@ -659,447 +542,4 @@ function _defaultErrorPredicate() {
  */
 function _defaultResponsePredicate() {
   return false
-}
-
-/**
- * Wraps the given function in a middleware recorder function.
- *
- * @private
- * @param {WebFrameworkShim} shim
- *  The shim used for this instrumentation.
- * @param {Function} middleware
- *  The middleware function to record.
- * @param {MiddlewareSpec} spec
- *  The spec describing the middleware.
- * @returns {Function} The middleware function wrapped in a recorder.
- */
-function _recordMiddleware(shim, middleware, spec) {
-  /**
-   *
-   */
-  function getRoute() {
-    let route = spec.route || '/'
-
-    if (shim.isFunction(route)) {
-      route = route()
-    }
-
-    if (route instanceof RegExp) {
-      route = '/' + route.source + '/'
-    } else if (shim.isArray(route)) {
-      route = route.join(',')
-    } else if (route[0] !== '/') {
-      route = '/' + route
-    }
-
-    return route
-  }
-
-  const typeDetails = MIDDLEWARE_TYPE_DETAILS[spec.type]
-  const name = spec.name || shim.getName(shim.getOriginal(middleware))
-  let metricName = shim._metrics.PREFIX + typeDetails.name
-  if (typeDetails.record) {
-    metricName = shim._metrics.MIDDLEWARE + metricName + name
-  }
-
-  /**
-   * @param route
-   */
-  function getSegmentName(route) {
-    let segmentName = metricName
-    if (typeDetails.path) {
-      segmentName += route
-    } else if (route.length > 1) {
-      segmentName += '/' + route
-    }
-
-    return segmentName
-  }
-
-  const isErrorWare = spec.type === MIDDLEWARE_TYPE_NAMES.ERRORWARE
-  const getReq = shim.isFunction(spec.req) ? spec.req : _makeGetReq(shim, spec.req)
-
-  assignCLMSymbol(shim, middleware)
-
-  return shim.record(
-    middleware,
-    spec.promise ? middlewareWithPromiseRecorder : middlewareWithCallbackRecorder
-  )
-
-  // TODO: let's please break these out
-  /**
-   * @param shim
-   * @param fn
-   * @param fnName
-   * @param args
-   */
-  function middlewareWithCallbackRecorder(shim, fn, fnName, args) {
-    const route = getRoute()
-
-    // Pull out the request object.
-    const req = getReq.call(this, shim, fn, fnName, args)
-
-    // Fetch the transaction information from that request.
-    const txInfo = _getTransactionInfo(shim, req)
-    if (!txInfo || !txInfo.transaction) {
-      shim.logger.debug(
-        { txInfo: txInfo },
-        'Could not get transaction info in %s (%s)',
-        route,
-        fnName
-      )
-      return null
-    }
-    txInfo.transaction.nameState.setPrefix(shim._metrics.FRAMEWORK)
-    txInfo.errorHandled |= isErrorWare
-
-    // Copy over route parameters onto the transaction root.
-    let params = shim.agent.config.high_security
-      ? null
-      : spec.params.call(this, shim, fn, fnName, args, req)
-
-    // Route parameters are handled here, query parameters are handled in lib/transaction/index.js#_markAsWeb as part of finalization
-    params = shim.prefixRouteParameters(params)
-
-    // Wrap up `next` and push on our name state if we find it. We only want to
-    // push the name state if there is a next so that we can safely remove it
-    // if context leaves this middleware.
-    let nextWrapper = null
-    if (shim.isFunction(spec.next)) {
-      const nextDetails = {
-        route,
-        wrapNext: spec.next,
-        isErrorWare,
-        isPromise: false,
-        appendPath: spec.appendPath
-      }
-
-      nextWrapper = _makeNextBinder(nextDetails, txInfo)
-    } else {
-      const nextIdx = shim.normalizeIndex(args.length, spec.next)
-      if (nextIdx !== null && args[nextIdx] instanceof Function) {
-        const nextDetails = {
-          route,
-          wrapNext: function wrapNext(s, f, n, _args, wrap) {
-            wrap(_args, nextIdx)
-          },
-          isErrorWare,
-          isPromise: false,
-          appendPath: spec.appendPath
-        }
-
-        nextWrapper = _makeNextBinder(nextDetails, txInfo)
-      }
-    }
-
-    // Append this middleware's mount point if it's not an errorware...
-    // (to avoid doubling up, a la 'WebTransaction/Expressjs/GET//test/test')
-    if (!isErrorWare && spec.appendPath) {
-      txInfo.transaction.nameState.appendPath(route, params)
-    }
-
-    // ...and possibly construct a recorder
-    let recorder = null
-    if (typeDetails.record) {
-      const stackPath = txInfo.transaction.nameState.getPath() || ''
-      recorder = _makeMiddlewareRecorder(shim, metricName + '/' + stackPath)
-    }
-
-    const segmentName = getSegmentName(route)
-
-    // Finally, return the segment descriptor.
-    return {
-      name: segmentName,
-      callback: nextWrapper,
-      parent: txInfo.segmentStack[txInfo.segmentStack.length - 1],
-      recorder: recorder,
-      parameters: params,
-      after: function afterExec(shim, _fn, _name, err) {
-        const errIsError = _isError(shim, err)
-        if (errIsError) {
-          _noticeError(shim, txInfo, err)
-        } else if (!nextWrapper && !isErrorWare && spec.appendPath) {
-          txInfo.transaction.nameState.popPath(route)
-        }
-        if (errIsError || !nextWrapper) {
-          txInfo.segmentStack.pop()
-        }
-      }
-    }
-  }
-
-  /**
-   * @param shim
-   * @param fn
-   * @param fnName
-   * @param args
-   */
-  function middlewareWithPromiseRecorder(shim, fn, fnName, args) {
-    const route = getRoute()
-
-    // Pull out the request object.
-    const req = getReq.call(this, shim, fn, fnName, args)
-
-    // Fetch the transaction information from that request.
-    const txInfo = _getTransactionInfo(shim, req)
-    if (!txInfo || !txInfo.transaction) {
-      shim.logger.debug(
-        { txInfo: txInfo },
-        'Could not get transaction info in %s (%s)',
-        route,
-        fnName
-      )
-      return null
-    }
-    txInfo.transaction.nameState.setPrefix(shim._metrics.FRAMEWORK)
-    txInfo.errorHandled |= isErrorWare
-
-    // Copy over route parameters onto the transaction root.
-    let params = shim.agent.config.high_security
-      ? null
-      : spec.params.call(this, shim, fn, fnName, args, req)
-
-    // Route parameters are handled here, query parameters are handled in lib/transaction/index.js#_markAsWeb as part of finalization
-    params = shim.prefixRouteParameters(params)
-
-    // Append this middleware's mount point and possibly construct a recorder.
-    if (spec.appendPath) {
-      txInfo.transaction.nameState.appendPath(route, params)
-    }
-    let recorder = null
-    if (typeDetails.record) {
-      const stackPath = txInfo.transaction.nameState.getPath() || ''
-      recorder = _makeMiddlewareRecorder(shim, metricName + '/' + stackPath)
-    }
-
-    // The next callback style can still apply to promise based
-    // middleware (e.g. koa).  In this case we would like to remove the
-    // path for the current executing middleware, then readd it once the
-    // next callback is done (either asynchronously or after the
-    // returned promise is resolved).
-    let nextWrapper = function pushSegment(shim, _fn, _name, segment) {
-      txInfo.segmentStack.push(segment)
-    }
-    if (shim.isFunction(spec.next)) {
-      const nextDetails = {
-        route,
-        wrapNext: spec.next,
-        isErrorWare,
-        isPromise: true,
-        appendPath: spec.appendPath
-      }
-      nextWrapper = _makeNextBinder(nextDetails, txInfo)
-    } else {
-      const nextIdx = shim.normalizeIndex(args.length, spec.next)
-      if (nextIdx !== null && args[nextIdx] instanceof Function) {
-        const nextDetails = {
-          route,
-          wrapNext: function wrapNext(s, f, n, _args, wrap) {
-            wrap(_args, nextIdx)
-          },
-          isErrorWare,
-          isPromise: true,
-          appendPath: spec.appendPath
-        }
-
-        nextWrapper = _makeNextBinder(nextDetails, txInfo)
-      }
-    }
-
-    const segmentName = getSegmentName(route)
-
-    // Finally, return the segment descriptor.
-    return {
-      name: segmentName,
-      parent: txInfo.segmentStack[txInfo.segmentStack.length - 1],
-      promise: spec.promise,
-      callback: nextWrapper,
-      recorder: recorder,
-      parameters: params,
-      after: function afterExec(shim, _fn, _name, err, result) {
-        if (shim._responsePredicate(args, result)) {
-          txInfo.transaction.nameState.freeze()
-        }
-        if (_isError(shim, err)) {
-          _noticeError(shim, txInfo, err)
-        } else {
-          txInfo.errorHandled = true
-
-          if (spec.appendPath) {
-            txInfo.transaction.nameState.popPath(route)
-          }
-        }
-        txInfo.segmentStack.pop()
-      }
-    }
-  }
-}
-
-/**
- * @param shim
- * @param req
- */
-function _makeGetReq(shim, req) {
-  return function getReqFromArgs(shim, fn, name, args) {
-    const reqIdx = shim.normalizeIndex(args.length, req)
-    if (reqIdx === null || !args[reqIdx]) {
-      shim.logger.debug('Can not find request parameter, not recording.')
-      return null
-    }
-    return args[reqIdx]
-  }
-}
-
-/**
- * @param nextDetails
- * @param txInfo
- */
-function _makeNextBinder(nextDetails, txInfo) {
-  return function bindNext(shim, fn, _name, segment, args) {
-    if (!segment) {
-      return
-    }
-    txInfo.segmentStack.push(segment)
-
-    nextDetails.wrapNext(shim, fn, _name, args, nextWrapper)
-
-    // Called from outside to wrap functions that could be called to continue
-    // to the next middleware
-    /**
-     * @param nodule
-     * @param property
-     * @param isFinal
-     */
-    function nextWrapper(nodule, property, isFinal) {
-      shim.wrap(nodule, property, function wrapper(shim, original) {
-        const parentSegment = segment || shim.getSegment()
-        return shim.bindSegment(function boundNext(err) {
-          // Only pop the stack if we didn't error. This way the transaction
-          // name is derived from the failing middleware.
-          if (_isError(shim, err)) {
-            _noticeError(shim, txInfo, err)
-          } else if (!isFinal && !nextDetails.isErrorWare && nextDetails.appendPath) {
-            segment.transaction.nameState.popPath(nextDetails.route)
-          }
-
-          // The next call does not signify the end of the segment
-          // calling next in the promise case.  Keep the segment on the
-          // stack and wait for its promise to be resolved to end it.
-          if (!nextDetails.isPromise) {
-            txInfo.segmentStack.pop()
-            segment.end()
-          }
-          const ret = original.apply(this, arguments)
-
-          if (nextDetails.isPromise && shim.isPromise(ret)) {
-            // After the next call has resolved, we should reinstate the
-            // segment responsible for calling next in case there is
-            // more work to do in that scope.
-            return ret.then(function onNextFinish(v) {
-              if (nextDetails.appendPath) {
-                segment.transaction.nameState.appendPath(nextDetails.route)
-              }
-
-              txInfo.segmentStack.push(segment)
-
-              return v
-            })
-          }
-
-          return ret
-        }, parentSegment) // Bind to parent.
-      })
-    }
-  }
-}
-
-/**
- * Retrieves the cached transaction information from the given object if it is
- * available.
- *
- * @private
- * @param {WebFrameworkShim}      shim  - The shim used for this instrumentation.
- * @param {http.IncomingMessage}  req   - The incoming request object.
- * @returns {object?} The transaction information if available, otherwise null.
- */
-function _getTransactionInfo(shim, req) {
-  try {
-    return req[symbols.transactionInfo] || null
-  } catch (e) {
-    shim.logger.debug(e, 'Failed to fetch transaction info from req')
-    return null
-  }
-}
-
-/**
- * Creates a recorder for middleware metrics.
- *
- * @private
- * @param shim
- * @param metricName
- * @param {string}  path    - The mounting path of the middleware.
- * @param {Segment} segment - The segment generated for this middleware.
- * @param {string}  scope   - The scope of the metric to record.
- */
-function _makeMiddlewareRecorder(shim, metricName) {
-  return function middlewareMetricRecorder(segment, scope) {
-    const duration = segment.getDurationInMillis()
-    const exclusive = segment.getExclusiveDurationInMillis()
-    const transaction = segment.transaction
-
-    if (scope) {
-      transaction.measure(metricName, scope, duration, exclusive)
-    }
-    transaction.measure(metricName, null, duration, exclusive)
-  }
-}
-
-/**
- * Adds the given error to the transaction information if it is actually an error.
- *
- * @private
- * @param {WebFrameworkShim} shim
- *  The shim used for this web framework.
- * @param {TransactionInfo} txInfo
- *  The transaction context information for the request.
- * @param {*} err
- *  The error to notice.
- */
-function _noticeError(shim, txInfo, err) {
-  txInfo.error = err
-  txInfo.errorHandled = false
-}
-
-/**
- * Determines if the given object is an error according to the shim.
- *
- * @private
- * @param {WebFrameworkShim} shim
- *  The shim used for this web framework.
- * @param {?*} err
- *  The object to check for error-ness.
- * @returns {boolean} True if the given object is an error according to the shim.
- */
-function _isError(shim, err) {
-  return err && shim._errorPredicate(err)
-}
-
-/**
- * Copy the keys expected from source to destination.
- *
- * @private
- * @param {object} destination
- *   The spec object receiving the expected values
- * @param {object} source
- *   The spec object the values are coming from
- */
-function _copyExpectedSpecParameters(destination, source) {
-  const keys = ['matchArity']
-
-  for (let i = 0; i < keys.length; ++i) {
-    const key = keys[i]
-    if (source[key] != null) {
-      destination[key] = source[key]
-    }
-  }
 }
